refactor: split outbox_handler.py into addressing and delivery sub-modules (#971)
Extract outbox addressing/dehydration concerns into outbox_addressing.py
and object validation/preparation concerns into outbox_delivery.py.
outbox_handler.py retains the emitter singleton, re-exports all symbols
from the two new modules for backward compatibility (monkeypatch paths),
_prepare_activity_object_for_delivery, and the two public coroutines.

Main handler drops from 560 lines to 262; new modules are 167 and 236 lines.
No behavior changes.
